Renamed client_weights.py to extract_client_weights.py
